model2/requirements.txt
torch>=2.0.0
transformers>=4.34.0
accelerate>=0.23.0
bitsandbytes>=0.41.0
unsloth>=2023.11.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0